Bug#35710213: mysql server 8.1.0 failed at Item_field::used_tables
The problem is in the function JOIN::destroy(), where we call
update_used_tables() to reinstate used tables information after
execution of a statement that has one or more const tables, and has thus
been subject to const table elimination. During this, we may encounter
a table, such as a special information schema, that has already been
closed and subsequently has had its table pointer (in class Table_ref)
set to nullptr.

The solution to this is to delay the refreshing of used table
information to the start of the next execution, just after tables have
been opened and we know that all table objects are in a proper state.

The performance impact of this seems to be negligible.

Also added some necessary code to prevent that uninitialized command
properties were copied in the case of as yet unprepared statements.

Change-Id: Ib44ab2cf1e61c3b188e7ceb000adb1445c48979a