Bug#35945223	Errors on Gtid_log_event::do_apply_event might refer an old GTID
Issue description
-----------------
When an event with GTID is processed its current GTID is checked for format,
permissions and other before it is assigned to THD->gtid_next attribute.
During those steps an error can be detected, for example, wrong format
or lack of permissions. Then, when the error is reported gtid_next
from a previous transaction is used as the current one is not yet
assigned to THD->gtid_next.

Analysis
--------
Errors are reported by rpl_rli_pdb.cc Slave_worker::do_report function
which uses current THD->gtid_next value.

Proposed solution
-----------------
Introduce a way to pass GTID to Slave_worker::do_report function
so that current GTID being processed can be used in the function
before it is assigned to THD->gtid_next value.

Change-Id: I052c98854781a61bdb4f41e603374a3ed8f8c800