Bug#35750771 Refactor Transporter and NdbSockets to handle lifecycle issues
Post-push addendum patch, fixing a regression causing slow disconnects.

When the connection over a transporter was 'lost', possibly due to
the other peer closing the socket, the TransporterRegistry logic
to gracefully take down the connection relied upon ::performReceive()
to be called on the failed transporter.  performReceive() would then
detect an error on that socket, resulting in start_disconnecting_trp()
to be called, which would initiate the disconnect process, and eventually
report the Transporter as DISCONNECTED.

Patch for Bug#35750771 introduced handling of the EPOLLHUP event.
When receiving that event we just unsubscribed on any further
event from the HUP'ed socket. Notably we did *not* set the TrpId
bit in the m_recv_transporters which had data to be received
from ::performReceive().

Thus performReceive() may not detect the failed socket and
start_disconnecting_trp().

Patch fix this issue by calling start_disconnecting_trp() immediately
when the EPOLLHUP event is received.

Another minor issue is also fixed, where we counted the EPOLLHUP event
as if a transporter had been added to the m_recv_transporters set.

Change-Id: I526887f1f39c74e3c4ae166a45853cb7f9f50096