WL#16107 Hypergraph optimizer core cost model
The hypergraph optimizer was developed without a clearly defined cost
model. A mixture of the of the old optimizer's cost model and new
uncalibrated cost constants has been used in an ad-hoc way to assign
costs to different access paths during optimization.

Some of the problems with the current cost model are:

1) There is no well-defined cost unit, i.e., it is unclear whether a
   cost of 1.0 is intended to represent a running time of one
   millisecond, one IO block transfer, or something else.

2) The cost model assigns cost to different operations in an
   inconsistent manner. For example, the current cost model for table
   scans is based entirely on the number of pages in the table, while
   the cost model for index range scans includes a cost for each
   processed row. This leads to inconsistent costs that do not reflect
   actual running times.

3) The cost of different operations has not been calibrated to reflect
   actual running times. Most cost constants such as the cost of
   comparisons during sorting, the cost of hashing rows during hash
   joins, and the cost of filtering rows has been assigned the same
   constant cost value.

This patch lays the groundwork for changing the hypergraph cost model
and includes the following:

1) A well-defined cost unit that reflects running time.

2) A consistent cost model for base table access paths: TABLE_SCAN,
   INDEX_SCAN, INDEX_RANGE_SCAN, REF, and EQ_REF.

3) Additional calibration of the cost models for SORT, HASH_JOIN and
   FILTER access paths.

Due to the scope and complexity of the cost model this patch only
represents a partial move to the new cost model in the sense that some
access paths and parts of the hypergraph optimizer still rely on the old
cost model. The cost constants in the old model have been adjusted for
the hypergraph optimizer in an attempt to make the two models somewhat
compatible.

We currently lack benchmarks that are comprehensive (cover all types of
queries and optimizations) and benchmarks that are representative of how
customers use MySQL which we could use to evaluate changes to the cost
model. Here is an overview of current benchmark results:

- TPC-H: No regressions, improvements to a few queries.
- TPC-DS: A few regressions and a few improvements. These queries are
  very complex so the combination of row estimation errors, bugs in the
  hypergraph optimizer, and the fact that the cost model is only
  partially updated all factor in.
- Cost benchmark made for this patch: Mostly improvements, a few
  regressions due to e.g. bugs in how the range optimizer produces row
  estimates.

This cost model patch has been constructed for the InnoDB storage
engine, and has only been calibrated for a data that resides in the
buffer pool (main memory). We leave it as future work to extend the cost
model to support multiple storage engines and different IO costs. Making
the cost model robust to the various states of optimizer statistics (and
the differences in statistics between storage engines) also remains.

Change-Id: Ic4ab2dc5b3c08c225256bc10657560f380d7ca44