Bug#35507763: Assertion `to->field_ptr() != from->field_ptr()' failed.
An assertion failed when inserting data into a table with a
zero-length column, such as CHAR(0) or BINARY(0).

The purpose of the assertion was to detect if a column was copied into
itself, but it could also be triggered when copying between adjacent
columns if one of the columns had zero length, since the two columns
could have the same position in the record buffer in that case. This
is not a problem, since copying zero bytes is a no-op.

Fixed by making the assertion less strict and only fail if it detects
that a non-zero number of bytes is copied from a source that is
identical to the target.

Change-Id: Ifc07b13a552c3ef6ce99bcc1d491afeef506e9c1
(cherry picked from commit d0463fd8b4657911589ecbf968eaad64983cd672)