Bug#35728352 Autotest testNdbApi -nCheckSlowCommit fails occasionally
Testcase was failing on multi-nodegroup clusters as it was killing
a data node which was not a participant in the transaction being
tested, and so the TC node failure handling code being tested
was not invoked.

In a single nodegroup cluster, all data nodes are participants
in any writing transaction.

Fix is to ensure that the node killed is both :
 - Not the TC node
 - A participant in the transaction.

Change-Id: Idf60a79918b648f5125a681204b9c168c8de37dc