Bug#35934223 [noclose] spurious failures in ndb.test_mgmd
The "Bug45495" test in testMgmd fails under gcov testing. The
test results show hundreds of lines of DEBUG log messages
being written while a fiile_exists() call times out.
This patch starts the mgmd servers without the --verbose flag,
so that the DEBUG level log messages are suppressed.

This patch is intended for both mysql-8.0 and mysql-trunk.
A separate patch for this bug addresses failures in other tests
and is intended for mysql-trunk only.

Change-Id: I904de0a1926f3a151ee21c194659b61b2d19aed2