Bug#33045412: ASSERTION `M_TYPELIB != NULLPTR' FAILED|SQL/ITEM.CC [2/2, bug fix]
Some queries that used the LEAD or LAG window functions on a column of
a SET or ENUM type hit an assertion during resolving. The same
assertion was hit in some queries using the LEAST or GREATEST function
on a SET or ENUM column. The reason was that Item_lead_lag and
Item_func_min_max didn't implement get_typelib().

Fixed by implementing the missing get_typelib() functions.

Change-Id: Ifd10ccdd65da1495201f18e89ef66efbd9426564