Bug#30529132 REVIEW OUTDATED LCPSCANFRAGWATCHDOG AND BUG24664 USING ERROR INSERT 10039
This patch adds an error insert mechanism to allow LCP fragment
scans to be stalled once they start (10055), or after scanning
at least one batch of rows (10056).
Additionally, a table id can be supplied to indicate that the
stall should happen on a particular table.
To supply the tableId the error insert 'extra' value is used.
This can be specified using e.g. the NdbRestarter test class.
The ndb_mgm ERROR command does not currently allow it to be
specified, but the Backup block already has a dump code
(DUMP 13003) that can be used to set the error and error2
codes in the Backup block.

Therefore :
  Stall next LCP fragment scan to start :
    ERROR 10055
    DUMP 13003 10055 0
  Stall next LCP fragment scan to start on tableid t
    ERROR 10055 <t>
    DUMP 13003 10055 <t>

  Stall next LCP fragment next-batch :
    ERROR 10056
    DUMP 13003 10056 0

  Stall next LCP fragment next-batch on tableid t
    ERROR 10056 <t>
    DUMP 13003 10056 <t>

Subsequent patch will use these error codes from existing
tests which are not currently getting LCP coverage.

Change-Id: I24cdceb69782b101953d8a94b0e5539d2eb91e31