Bug#35991384: Mysqld crash from Query_block::optimize() Assertion `join == nullptr'
An assertion failure was seen when killing a query while it was
evaluating an uncorrelated subquery containing a hash join during
optimization.

This happened because
HashJoinIterator::ReadNextJoinedRowFromHashTable() stopped the join
and returned an error code once it detected that the query had been
killed, but without setting the error flag in the THD. Since returned
error codes cannot be propagated through all interfaces (this
particular one was lost in a val_int()), and callers higher up
therefore need to be informed via thd->is_error(), which wasn't set,
the optimization continued without noticing that the optimization of a
subquery had not completed, and assertions failed because some of the
optimizer data structures were in a bad state.

Fixed by making HashJoinIterator::ReadNextJoinedRowFromHashTable() set
the error flag in the THD when it stops a join midway because the
query has been killed.

Change-Id: Idd5bc99eec7fe2a65861f5bfa5317cc8a8f7e2e3