Bug#32610988: GR MEMBER TRIES TO CONNECT TO SERVER EVEN IF THE SERVER IS OUT OF THE GROUP.
Problem
=======================

Consider two member M1 and M2 are in group with MySQL
communication stack with M1 using u1 as recovery user and
M2 using u2 as recovery user both users u1 and u2 exists on
M1 and M2 with necessary privileges. Join a new member M3
to the group using 'u2' as recovery user. M3 has only user
'u2'. It doesn't have user 'u1'. The start GR on M2 fails
as M1 can't connect to M2. Hence a new view_id is generated
in group stating the members in group as M1 and M2. But
still after this M1 continues to try to connect to M3 and
hence logs Error in establishing mysql connection to the
Error log and M3 logs access denied error for connection
from M1.

As per the discussions, the behaviour exists even with XCom
stack. It is just visible from MySQL stack from the server
Error log

Analysis and Fix
======================

By design, XCom stores the last three known configurations.
As such, it also keeps references to the physical
connections that are shared among all past and present
configurations. It takes advantage of that fact in
order to quickly reconnect nodes that rejoin the group,
explicitly or implicitly, and that were already present
in any of those configurations.

The side effect of this, is that we might keep connecting a
node that is currently not in the group.

A simple solution for this, and the solution that is in
this patch, is to simply inhibit error logging if the node
is not in the current configuration, in order to avoid
false negative problems, leading a DBA or an operator
to think that they have a problem in the system.

Change-Id: I8c366dab28baf0c29d96309bfeeaf626313a5896