Bug#36002814 Test failure in testNodeRestart -n LCPScanFragWatchdog T2
The Qmgr GSN_ISOLATE_ORD handling was modified to handle the larger
node bitmap size necessary for supporting up to 144 data nodes.

For compatibility QMGR uses the node version of the sending node to
determine whether the incoming signal has its node bitmap in a long
section or inline in the signal.

The senderRef member of the incoming signal is used which is set
by the signal originator.

However in the context of ISOLATE_ORD, the original sender may be
shutdown when ISOLATE_ORD is processed, in which case their node
version may have been reset to zero, causing the inline bitmap
path to be taken, resulting in incorrect processing.

The signal handler is changed to use the presence of a long section
to decide whether the incoming signal uses a long section to
represent nodes to isolate or not.

Change-Id: I1efe6c98c3f342770462d81ae00395b5a2eec7d7