Bug#36004838 An overloaded transporter fails to yield the send-thread
A transporter may be set in an 'overloaded' state if a send attempt
failed to send anything on it. That could e.g be to the OS-internal
send buffers being filled up, possibly due to the receiver side not
consuming fast enough / having full buffers as well.

The handling of such overload is to let the send-thread taking
a short 200us yield from the CPU if there is no other send work
to be found.

Right before the send-threads are yielding the CPU the yield method
will check the callback method check_available_send_data() for any
late arrived transporters having more data to send. That is effectively
implemented by the send-threads having the flag 'bool m_more_trps' which
is set every time a new transporter is entered into the queue with
insert_trp(). It is also cleared if get_trp() could not find a TrpId
to be returned.

The special case that get_trp() had found only a delayed transporter
is handled in handle_send_trp(). Here we cleared the 'm_more_trps'
if we found a delay, and handle_send_thread was part of a send-thread
(as opposed to an assist-send)
Note that failing to clear the m_more_trps would result in the
send-thread later not being able to yield the CPU.

For the 'delayed-send with overload'-case however, handle_send
did an early 'return false', without ever entering the code section
where m_more_trps was cleared for a send-thread. This prevented
the CPU yield in the overload situation.

Patch is to combine the check for overload with the general check
for delay to be canceled or m_more_trps to be cleared.

Change-Id: Ied5cca628e10715510d88350140c7f4ce1a17f37