Bug#35931702 When binlog_format=MIXED, writeset dependency tracking is broken because statement part is ignored
When Mixed mode is used for binary logging, and a transaction exists
that contains a part logged as a statement (a safe query) and a part
that is logged in row mode (unsafe query) then only partial write set
information exists for the query.

Using the write set information for conflict detection could lead to
false negatives, and for that reason we limit the usage of write sets
for conflict checks to only row based logging.

Change-Id: I55f9094d96afc69b5589c4fa33488a4019d7da6d