Bug#36005903 mgmapi failure to parse bindadress
Problem:
Trying to start ndb_mgmd with --bind-address=localhost fails with error
"Illegal bind address". This error is returned from the MgmApi when
parsing the bind address to split it into a host and port part.

  $> ndb_mgmd --bind-address=localhost
  MySQL Cluster Management Server mysql-8.3.0 ndb-8.3.0
  2023-11-13 15:25:05 [MgmtSrvr] ERROR    -- Illegal bind address

Analysis:
The portnumber is parsed using strtol() and the errno value is examined.
The empty string cause EINVAL and thus parsing is handled as failure.

Solution:
Using errno is known to be fragile and is not really neccessary to
sufficently parse the number. Remove usage of errno in order to accept
the valid bind address "localhost" as well as more or less any string on
the form "<hostname>:<number>". Max and min value checks are left in
place.

Also unrelated fix of missing argument names in doxygen comment.

Change-Id: I177164fdc889ec4a1f5568e7064190babee6dd13