Bug#36054923 TcpPortPool on solaris gives out non-unique ports
On Solaris the "router_require" tests fails all the time
as the TcpPortPool gives out port's that are unique.

The TcpPortPool allows to reserve TCP port ranges between multiple
applications.

Using multiple TcpPortPools in the same applications fails on Solaris as
taking a file-lock that's own

- by another process is exclusive
- by same process is non-exclusive

On all other Unixes, it is exclusive in both cases.

Change
======

- refactored the file-descriptor/handle handling into FileHandle class
- track unique-ids owned by the process to make file-locks exclusive
  within the same process

Change-Id: Ieeef9a2066f150a9be403a135a332cc90f7e145e