Bug#36053908 Windows ndb_sign_keys assert(!running()) fails.
For debug builds, if ndb_sign_keys fails to start a co process when for
example using the --CA-tool option it fails on assert(!running()) in
NdbProcess destructor.

For debug builds the destructor of NdbProcess checks that the process is
not running by comparing process handle to INVALID_HANDLE_VALUE aliased
as InvalidHandle.

The used Windows function CreateProcess zeroes out the process and
thread handles on failure, not using the expected INVALID_HANDLE_VALUE.

Change sets the handle values when CreateProcess fails.

Change-Id: I4d5d2b40de0b589601411cc342fb473dfab295bf