BUG#35940509: "memory/group_rpl/Gcs_message_data::m_buffer" reports negative values in SPM
Group Replication (GR) rely on the Group Communication Service (GCS)
to exchange messages with the group and be informed of the group
membership.
When there is event from GCS, for example a message or membership
update, the GCS delivery thread will call the respective handler on
GR so that GR pursuits the required actions. The GCS delivery thread
is not initialized with a MySQL context. Some operations done on the
GR event handlers do require a MySQL context to be performed on
debug builds, the context is required for the debug information
provided by the performance schema instrumentation framework.
The operations that do require a MySQL context on debug builds are:
 * GR member actions load and store;
 * asynchronous replication channels failover configuration load and
   store;
 * debug flags for conditional execution.
To allow these operations, a temporary context was created, that is,
they were surrounded by:
```
  my_thread_init();
  ...
  my_thread_end();
```
These operations only happened on single-primary mode.

On 8.X.0 it was noticed that this temporary MySQL context was
causing side-effects, more precisely it was disabling the memory
consumption tracking inside GCS.

To solve the issue on 8.X.0+ the following changes were done:
1) The temporary MySQL context creation was removed from the GR
   event handlers.
2) A single MySQL context is created for the entire lifetime of
   the GCS delivery thread if GCS is built together with MySQL
   server.
3) Moved the `gms_listener_test` test to run on a thread which
   already has a MySQL context. The original `gms_listener_test`
   was also creating a temporary context.

Change-Id: I59c481cde76bc3995e585eeb130b53676c25bd80