Bug#33725447: A bug in handler.cc cause mysql-server crash               when input some sql statement
When a shared materialized table for a cte is marked
for re-materialization because it is a lateral derived
table and the counters for the invalidators do not
match for the second reference to the cte, it clears
all references to the cte which resets the cursors
positioned for the table which materialzed the cte
table leading to problems when reading from that table.
More detailed analysis in the bug page and the test
file (w.r.t the failing query).
The fix is to not re-materialize the cte table when
we are using the shared materialized derived table.
We also update the invalidator counters while at
it.

Change-Id: I5f9daeab03fee93354c0253fd55865f1a6a1e32c