BUG#34951273 : Performance of scanning data_lock_waits worse than expected with read-only trx
The performance_schema.data_locks and performance_schema.data_lock_waits
tables are not real InnoDB tables, but rather a way for the user to see
the state of in-memory data structures used by InnoDB to track data
locks - it is constructed on the fly by special iterators which go
through the in-memory data structures and populate the row data with
information extracted from them.
This is done in batches to not keep the global latches on trx_sys and
log_sys for too long.
The batch was defined in terms of a range on printable trx_id values,
where each range spanned exactly 256 numbers. A single batch processed
all locks of all transactions from a given range, and then prepared the
range for the next batch, setting its lower end to the minimal not yet
processed id.
While this approach ensured no batch had more than 256 transactions to
process (under assumption that ids are unique), it had an opposite
problem: if ids were spread far appart, a batch could contain just a
few, perhaps just one transaction. This could lead to Omega(N) batches,
where each had to take the latches, scan O(N) transactions to find these
falling into the range, then release latches.
One scenario in which this was common, was for read transactions, which
had no "real" trx->id assigned yet, for which we used a fake, but unique
identifier derived from trx_t object's address. As sizeof(trx_t)>256, no
two such transactions could be processed in the same batch.

The solution proposed in this patch is to define the range of the batch
so that it contains 256 transactions, as opposed to 256 numbers. This is
achieved by scanning the list of not yet processed transactions, and
picking the 256-th smallest of them as the end of the range. This way
each batch contains as many elements as possible, and thus the number of
batches is minimized.
There are several ways to identify 256-th smallest element.
The simplest would be to use n_th_element, but this requires Omega(N)
memory.
To upperbound the memory usage of the scan, we keep track of the 256
smallest elements when iterating over the list, arranging them into a
heap so that the largest of them can be easily identified, or replaced
with the visited element if it should be among those 256 smallest.
This is achieved by introducing Max_of_n_smallest<256,uint64_t> class.
This change shouldn't affect expected time complexity if the sequences
are random or increasing.
The worst case would be if the sequence is decreasing, in which case
each visited element should be inserted into the heap of 256 smallest
numbers, which would cause lg256=8 operations per item.
So, at worst this implementation can be 8 times slower than the eariler
one for a single batch, but hopefully it makes it up in reducing the
number of batches significantly.
In the artificial test of iterating over 18 000 read only transactions
it reduced the time from 5 seconds to 0.

Change-Id: I4cc28fc869f323a7b38adc4cfb88cd5be0fa5d32