Bug#35856247: Assertion Item_ref::walk at sql/item.h
Assertion failures were seen in some prepared statements using GROUP
BY WITH ROLLUP and lateral derived tables when the hypergraph
optimizer was enabled.

The problem is that the items in the Query_block::base_ref_items array
aren't properly restored at the end of execution, so that the next
execution of the statement sees references to fields in temporary
tables that are not preserved between executions.

There is code in JOIN::cleanup() to restore the original
base_ref_items array. However, it interacts poorly with the regular
rollback of changes to the item tree performed by
THD::rollback_item_tree_changes(). What happens is:

1) JOIN::refresh_base_slice() adds rollup group wrappers and rollup
sum wrappers, if needed.

2) Later in optimization, the hypergraph optimizer replaces the child
of a rollup group wrapper with a reference to a temporary table field.

3) At the end of the execution, JOIN::cleanup() replaces the rollup
wrappers in base_ref_items with the unwrapped items. Since the child
of the wrapper is still the temporary table field, it is the temporary
table field that gets stored in the array.

4) Later, still at the end of the execution,
THD::rollback_item_tree_changes() replaces the temporary table field
reference in the rollup wrapper with the original item, but too late
for JOIN::cleanup() to see it and put it back into base_ref_items.

5) The prepared statement is re-executed, and it fails when trying to
access the dangling temporary table field reference.

The bug is fixed by storing a copy of the original base ref items at
the end of preparation (in Query_block::save_properties()), and
restoring them at the beginning of each execution (in
Query_block::restore_cmd_properties()).

The code added in bug#14105951, bug#33349994 and bug#33847722 to
restore base_ref_items is removed, because it is no longer necessary
after these changes.

Change-Id: I33e6c0afc22e0e83543413cf407d4c2732206efd