Bug#35997316 Hypergraph: Incomplete and inconvenient optimizer trace
The existing mechanism for optimizer trace from the Hypergraph optimizer
passes a pointer to a std::string to every function that may need to
write trace. This mechanism has some shortcomings:

* The pointer to trace_str must be passed to every function that may
  directly or indirectly add trace. This is inconvenient, and, in case
  of overloaded operators, impossible. This means that trace is
  missing in some places, because some function higher up on the stack
  does not take the trace_str pointer as a parameter.

* std::string lacks the '<<' formatting operators of std::ostream.

* Appending to a std::string may be inefficient, as it may require
  allocating and copying into a new buffer of consecutive memory.

This commit therefore adds a class 'UnstructuredTrace' that provides
an 'ostream' trace sink, and attaches an instance of that class to
Opt_trace_contex. The latter is available via THD, and is thus readily
available in most parts of the code.

Change-Id: I463db0fe7e1b50ef33feac5308b22c9c8eda88c0