Bug#35750394 MultiTransporter should follow the DISCONNECT 'protocol'
Post push addendum patch, remove multiTransporter locks
around update of Transporter::m_connected.

Prior to fixing bug Bug#35750394, we had to check the
Transporter::isConnected() state for members of multiTransporters.
That was needed as the performState[] was managed on node level,
not transporter level, prior to these patches. Thus checking
the Transporter::isConnected state directly was the only way to
know whether a transporter was connected or not.

-> That also introduced the need for any changes to the
transporter::m_coonnected flag to be protected with
a lockMultiTransporters().

As we now follow the protocol, even for multi transporter
member, there are no specific isConnected() checks for these.
Thus the need for protecting any such updates for
multiTransporters should be obsolete as well.

Note that the update of m_connected to 'false' in
Transporter::doDisconnect() was never lock protected ...!

Change-Id: If7d1e0ff3c1c8818f36c8155a29471f553982694