Bug#35940325 Add support for PGO build of MySQL server on Windows
Problem:
When attempting to build the instrumented version of the MySQL server,
(i.e. with the -DFPROFILE_USE=1 CMake option), the size of the .obj files
produced by the Visual Studio 2019 C++ compiler have grown to the point
that the linker error LNK1248 is encountered while building the sql_main
library.

Although it would be possible to work around this issue by excluding
individual files from the PGO build (by undoing the use of the /GL
compiler option on those files) the number of files that would be affected
is becoming unwieldy and could potentially affect the overall optimization.

Solution:
The /Z7 compiler option has been used up to now. This option embeds debug
symbol information in .obj files, and can be replaced by the /Zi option
that stored debug symbol information separately from the .obj files, in
a .pdb file.

Using the /Zi option (for PGO builds only) reduces the size of the
.obj files sufficiently so that it is no longer necessary to exclude any
library or individual file from using the /GL PGO compiler option.

Change-Id: I84d6fad32310bf9dddfd9d93e5fe0d2b9b6ae0ab