Bug #29623286 Refactor HandlerList as a handler [2/7]
The Logger class internally contains a LogHandler list which
contains all of the log handlers which will be given the opportunity
to log each event.

This functionality is extracted into a class which implements the
LogHandler interface. This allows alternate implementations to be
used, and a pipeline of log handlers to be defined.

Mutex changes:
 - m_handler_mutex is renamed to m_handler_creation_mutex
   This guards against multiple threads creating 'singleton'
   handler objects using the Logger factory methods

 - m_listMutex is added, which protects against concurrent list
   modifications and list iteration.

 - m_mutex is renamed to m_log_mutex and protects against concurrent
calls to the Logger API functions

Where multiple mutexes are taken together the order should be e.g.
  1- m_handler_creation_mutex : Protects against concurrent singleton
  changes
  2- m_log_mutex : Protects against concurrent Logger API usage
  3- m_list_mutex   : Protects against concurrent LogHandler list
changes

Change-Id: I201aa95fd9afae8ccf57dc2f172de6e15689ad31