Bug#35855294 Inconsistent results of SELECT statement with EXISTS clause - part II: IN/ALL/ANY [noclose]
This patch is a follow-up on the first patch for this bug issue: we now tackle

  w-expr IN/ALL/ANY correlated subquery

where w-expr is an expression directly containing a window function
that needs a frame buffer. As evaluation of the IN/ALL/ANY predicate
will depend on both

        a) the results of a window function, and
        b) a correlated subquery, i.e. its value must be re-evaluated
           for every value of the left side in sync with reading the
           input result set

we can only evaluate the subquery for the IN/ALL/ANY predicate once we
have the window function's value. Therefore, any outer references that
are needed for evaluating the subquery must also go to the windowing
output tmp file and frame buffer.  This is handled with an extension
in split_sum_func2.

Next, to prepare for evaluation at window processing time,
CreateFramebufferTable/ReplaceMaterializedItems must make sure that
such outer references are also replaced to point to the windowing
output file. This is done by the new function
possibly_outerize_replacement called from ReplaceMaterializedItems.

The previously added test in window_functions_bugs.test giving wrong
result due to an outer reference pointing to the windowing input now
works as expected.

Change-Id: If576b932048900e94368b9c60018aae52f815c12