Bug#35993300 Server assertion crash after upgrade on 8.0.30
A user might disable redo log by using
ALTER INSTANCE DISABLE INNODB REDO_LOG
and then SHUTDOWN the server, which (unless innodb_fast_shutdown=2)
will write all dirty pages from Buffer Pool back to tablespaces.
Because this will mean redo log is not needed to recover the
tablespaces, the instance will be considered "crash safe" even though
the redo logging is disabled.
This is why we have two separate flags in the redo log header:
LOG_HEADER_FLAG_NO_LOGGING
LOG_HEADER_FLAG_CRASH_UNSAFE
The first one would still be set in such scenario, but the second would
get cleared during SHUTDOWN.
Now, the interesting scenario is what happens if after such SHUTDOWN the
user tries to upgrade from 8.0.29 or ealier, to 8.0.30 or later.
The 8.0.30 introduced a new layout of redo log files, thus it tries to
upgrade the redo log files. This is possible if the redo log is
logically empty, which it is in this case.
The upgrade logic tries to delete old files and create new empty files.
The only problem is that there is a piece of logic in the middle of
log_sys_init logic, which seeing the LOG_HEADER_FLAG_NO_LOGGING infers
that redo logging should be disabled, so calls mtr_t::Logging::disable
which in turn concludes that this will make the instance not crash safe
so we should set the LOG_HEADER_FLAG_CRASH_UNSAFE flag in case we crash.
This makes some sense. However the logic of writing flags to header
assumes the new, 8.0.30+ specific redo log format, and in particular
tries to set LOG_HEADER_FLAG_FILE_FULL flag which isn't even defined in
8.0.29, which leads to mismatch between the expected set of flags, and
existing set of flags on disc and assertion failure.
The root cause here is that we try to write to legacy files, which we
shouldn't really try to do - we should treat them as read only, as the
logic to manipulate files in old format is already removed.
All we can and should do is to validate they are logically empty and
remove them ASAP, and recreate them using the new format.

This fix defers the logic of disabling the redo log (and thus updating
the flags in header) after the files are recreated, recovery is finished
and we start writing to the redo log.
It also fixes a smaller bug which was calling log_start() twice in case
of upgrade. Thankfully it turned out to be almost idempotent by
accident, but it could lead to some issues with Link_buf being
initialized twice or other problems in future.

Change-Id: I5e4609a3df000e4bae611f8cf23f87f9bee9da31